+++ HaselTweaks/Interfaces/ITweak.cs
@@ -4,6 +4,7 @@ public interface ITweak : IDisposable
 {
     string InternalName { get; }
     TweakStatus Status { get; set; }
+    bool IsObsolete { get; set; }
     void OnEnable();
     void OnDisable();
 }